requirements.txt
langchain
langchain-community
openai
flask
python-dotenv